rs_merkle
Struct PartialTree

pub struct PartialTree<T: Hasher> { /* private fields */ }

Partial tree represents a part of the original tree that is enough to calculate the root. Used in to extract the root in a merkle proof, to apply diff to a tree or to merge multiple trees into one.

It is a rare case when you need to use this struct on it’s own. It’s mostly used inside MerkleTree and MerkleProof

Implementations

impl<T: Hasher> PartialTree<T>

pub fn new() -> Self

Takes leaves (item hashes) as an argument and build a Merkle Tree from them. Since it’s a partial tree, hashes must be accompanied by their index in the original tree.

pub fn from_leaves(leaves: &[T::Hash]) -> Result<Self, Error>

This is a helper function to build a full tree from a full set of leaves without any helper indices

pub fn build( partial_layers: Vec<Vec<(usize, T::Hash)>>, depth: usize, ) -> Result<Self, Error>

pub fn depth(&self) -> usize

Returns how many layers there is between leaves and the root

pub fn root(&self) -> Option<&T::Hash>

Return the root of the tree

pub fn contains(&self, layer_index: usize, node_index: usize) -> bool

pub fn merge_unverified(&mut self, other: Self)

Consumes other partial tree into itself, replacing any conflicting nodes with nodes from other in the process. Doesn’t rehash the nodes, so the integrity of the result is not verified. It gives an advantage in speed, but should be used only if the integrity of the tree can’t be broken, for example, it is used in the .commit method of the MerkleTree, since both partial trees are essentially constructed in place and there’s no need to verify integrity of the result.

pub fn layer_nodes(&self) -> Vec<Vec<T::Hash>>

pub fn layers(&self) -> &[Vec<(usize, T::Hash)>]

Returns partial tree layers

pub fn clear(&mut self)

Clears all elements in the ree